+++ CHANGELOG.md
@@ -7,6 +7,9 @@ this project uses date-based 'snapshot' version identifiers.
 
 ## [Unreleased]
 
+### Added
+- Variable `textfiledir`
+
 ### Changed
 - Documentation improvements